+++ app/src/main/java/com/bestapp/rice/ui/mettinglist/MeetingListViewModel.kt
@@ -1,4 +1,4 @@
-package com.bestapp.rice.ui.meetingmanagement
+package com.bestapp.rice.ui.mettinglist
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
@@ -12,16 +12,16 @@ import kotlinx.coroutines.launch
 import javax.inject.Inject
 
 @HiltViewModel
-class MeetingManagementViewModel @Inject constructor(
+class MeetingListViewModel @Inject constructor(
     private val appSettingRepository: AppSettingRepository,
     private val userRepository: UserRepository,
     private val meetingRepository: MeetingRepository,
 ) : ViewModel() {
 
-    private val _meetingManagementUiState =
-        MutableStateFlow<List<MeetingManagementUiState>>(emptyList())
-    val meetingManagementUiState: StateFlow<List<MeetingManagementUiState>>
-        get() = _meetingManagementUiState
+    private val _meetingListUiState =
+        MutableStateFlow<List<MeetingListUiState>>(emptyList())
+    val meetingListUiState: StateFlow<List<MeetingListUiState>>
+        get() = _meetingListUiState
 
     fun getMeetingByUserDocumentID() = viewModelScope.launch {
         appSettingRepository.userPreferencesFlow.collect { userDocumentID ->
@@ -33,7 +33,7 @@ class MeetingManagementViewModel @Inject constructor(
                 }
             )
 
-            _meetingManagementUiState.value = meetings.map {
+            _meetingListUiState.value = meetings.map {
                 // TODO: UserRepository 내부에 메서드 추가후 연동, fun getReviewState(userDocumentID: String): Boolean
                 val isDoneReview = false // userRepository.getReviewState(userDocumentID)
