+++ app/src/main/res/layout/fragment_meeting_management.xml
@@ -1,134 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
-
-<androidx.constraintlayout.widget.ConstraintLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.meetingmanagement.MeetingManagementFragment">
+    tools:context=".ui.mettinglist.MeetingListFragment">
 
-    <com.google.android.material.appbar.AppBarLayout
-        android:id="@+id/abl"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent">
-
-        <com.google.android.material.appbar.MaterialToolbar
-            android:id="@+id/mt"
-            style="@style/AppBar"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:minHeight="?attr/actionBarSize"
-            app:navigationIcon="@drawable/baseline_arrow_back_24"
-            app:title="내 모임 보기"
-            app:titleCentered="true"/>
-
-    </com.google.android.material.appbar.AppBarLayout>
-
-    <com.google.android.material.card.MaterialCardView
-        android:id="@+id/cv_enable_meetings"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_margin="@dimen/default_margin20"
-        android:backgroundTint="@color/white"
-        app:cardCornerRadius="@dimen/default_margin16"
-        app:layout_constraintTop_toBottomOf="@+id/abl"
-        app:layout_constraintStart_toStartOf="parent">
-
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_margin="@dimen/default_margin12"
-            android:orientation="vertical">
-
-            <TextView
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                style="@style/body_title16"
-                android:text="참여중인 모임"/>
-
-            <include
-                android:id="@+id/item_enable_meeting1"
-                layout="@layout/item_my_meeting"
-                android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                android:layout_marginTop="@dimen/default_margin8"/>
-
-            <include
-                android:id="@+id/item_enable_meeting2"
-                layout="@layout/item_my_meeting"
-                android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                android:layout_marginTop="@dimen/default_margin8"/>
-            <include
-                android:id="@+id/item_enable_meeting3"
-                layout="@layout/item_my_meeting"
-                android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                android:layout_marginTop="@dimen/default_margin8"/>
-
-        </LinearLayout>
-
-    </com.google.android.material.card.MaterialCardView>
-
-    <com.google.android.material.divider.MaterialDivider
-        android:id="@+id/md_line"
-        android:layout_width="match_parent"
-        android:layout_height="1dp"
-        android:layout_marginHorizontal="@dimen/default_margin20"
-        android:layout_marginTop="@dimen/default_margin12"
-        app:layout_constraintTop_toBottomOf="@+id/cv_enable_meetings"
-        app:layout_constraintStart_toStartOf="parent"/>
-
-    <com.google.android.material.card.MaterialCardView
-        android:id="@+id/cv_disable_meetings"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="@dimen/default_margin12"
-        android:layout_marginHorizontal="@dimen/default_margin20"
-        android:backgroundTint="@color/white"
-        app:cardCornerRadius="@dimen/default_margin16"
-        app:layout_constraintTop_toBottomOf="@+id/md_line"
-        app:layout_constraintStart_toStartOf="parent">
-
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_margin="@dimen/default_margin12"
-            android:orientation="vertical">
-
-            <TextView
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                style="@style/body_title16"
-                android:text="지난 모임"/>
-
-            <include
-                android:id="@+id/item_disable_meeting1"
-                layout="@layout/item_my_meeting"
-                android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                android:layout_marginTop="@dimen/default_margin8"/>
-
-            <include
-                android:id="@+id/item_disable_meeting2"
-                layout="@layout/item_my_meeting"
-                android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                android:layout_marginTop="@dimen/default_margin8"/>
-
-            <include
-                android:id="@+id/item_disable_meeting3"
-                layout="@layout/item_my_meeting"
-                android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                android:layout_marginTop="@dimen/default_margin8"/>
-
-        </LinearLayout>
-
-    </com.google.android.material.card.MaterialCardView>
 
 </androidx.constraintlayout.widget.ConstraintLayout>
